Implement COMMAND and ECHO and have the server respond.
This now responds favorably on benchmarks. Very unscientific, just
running on my laptop, but with logs on this is 6x slower than redis on
an echo benchmark, and with logs off it's 5% faster! So, this indicates
that I'm in the ballpark of a reasonable design.

I think this could go a lot faster later on, because this isn't nearly
close to saturating either the network or memory bandwidth (it's 5 MB/s
right now, while memory bandwidth is > 1 Gb/s and my network is 1 Gb/s).

This gives me room to explore, if I decide to go into "make go brrrrr"
mode, or I can explore the other areas of DBs.