Create in-memory storage and implement GET and SET commands (#1)
This creates a basic in-memory storage engine and wires it up to the GET
and SET commands. To avoid passing each new object all the way down the
tree, this introduces the `Context` struct which will be used to provide
necessary information throughout the call stack (likely additions
include configuration, transaction-related info, etc.).

On my laptop (not a good benchmark, but what I have until I get some VMs
provisioned), single-core is roughly 0.75x the throughput of redis and
multi-core is 2x the throughput of redis. This is a good launching off
point for profiling it and figuring out what's slowing it down and then
making some good gains.