Add command-line configuration options (#11)
This creates preliminary command-line configuration options and plumbs
it to pass these through to their consumers. The initial options are for
configuring number of worker threads, size of a queue, bind address, and
storage base path, and more will come as needed.

These allow easy tuning for benchmarks and optimizations.

More should be added based on #7 , but these are what's available for
what's implemented right now. In the future, as we add features, the
corresponding flags should be added.